Move get_shared_tables to data_sharing module and fix reference to DatasetService
