+++ backend/dataall/modules/datasets/api/table/queries.py
@@ -1,6 +1,6 @@
 from dataall.base.api import gql
 from dataall.modules.datasets.api.table.input_types import DatasetTableFilter
-from dataall.modules.datasets.api.table.resolvers import get_table, list_shared_tables_by_env_dataset, preview
+from dataall.modules.datasets.api.table.resolvers import get_table, preview
 from dataall.modules.datasets.api.table.types import (
     DatasetTable,
     DatasetTableSearchResult,
@@ -35,14 +35,4 @@ previewTable = gql.QueryField(
     args=[gql.Argument(name='tableUri', type=gql.NonNullableType(gql.String))],
     resolver=preview,
     type=gql.Ref('QueryPreviewResult'),
-)
-
-getSharedDatasetTables = gql.QueryField(
-    name='getSharedDatasetTables',
-    args=[
-        gql.Argument(name='datasetUri', type=gql.NonNullableType(gql.String)),
-        gql.Argument(name='envUri', type=gql.NonNullableType(gql.String)),
-    ],
-    type=gql.ArrayType(gql.Ref('SharedDatasetTableItem')),
-    resolver=list_shared_tables_by_env_dataset,
 )
